Use InternalRedirect for all plugin redirects
The attribute self.uri_fmt is used both with HTTPRedirect and
InternalRedirect calls. In the new plugin structure, it is
necessary to make all redirects uniform, since InternalRedirect
is relative to current path (i.e. /plugins/kimchi) and
HTTPRedirect is not. This patch converts all plugin redirects
to use InternalRedirect, since it is less costly.

Signed-off-by: Lucio Correia <luciojhc@linux.vnet.ibm.com>
Signed-off-by: Gustavo Y. Ribeiro <gyr@linux.vnet.ibm.com>