Moving JS libraries to new folders and added license files
This commit updates jquery and jquery-ui to the latest compatible versions
(v1.11.3 and v1.11.4) and adds all the JS libraries that are required in order
to run and develop the new-ui.
Updated the old-ui login.html.tmpl and wok-ui.html.tmpl to point to correct Js
path otherwise the CherryPy would throw a 404 error when loading the pages.
Moved all libs/themes files to libs/jquery-ui/themes/ and added a copy of
jquery-ui license to this folder.
All the new libraries where added to ui/libs folder and configure.ac was
properly updated to read the new Makefiles.

*New-ui related:
   *Removed logo.ico and updated with favicon.ico in high resolution
    (name convention for IE9 compatibility). This change can be spotted in
    the browser tab bar with the old-ui.
   *Updated opensans.css to point to local font family when the user already
    has it installed.

Signed-off-by: samhenri <samuel.guimaraes@eldorado.org.br>